Refactoring funcs for bot requests and logs

+++ bot/bot_requests/enrollments.py
@@ -7,22 +7,32 @@ def create_enrollment(base_domain, token, lesson_id):
     """
     Creates an enrollment for the given lesson.
     """
+    if not token:
+        return {
+            'status': 'error',
+            'message': 'Token is required'
+        }
+
     url = base_domain + "/enrollments/"
     data = {
         "lesson": lesson_id,
     }
     headers = {
-        "Authorization": "Bearer " + token
+        "Authorization": "Bearer " + token,
     }
     response = requests.post(
-        url,
+        url=url,
         data=data,
         headers=headers,
     )
+
     logging.info(f"""
+    Creates an enrollment for the given lesson.
+    
     Sending request to {response.url}
     Response: {response.text}
     """)
+
     if response.status_code == 201:
         return {
             "status": "success",
@@ -39,18 +49,28 @@ def get_all_enrollments(base_domain, token):
     """
     Gets all enrollments.
     """
+    if not token:
+        return {
+            'status': 'error',
+            'message': 'Token is required'
+        }
+
     url = base_domain + "/enrollments"
     headers = {
         "Authorization": "Bearer " + token
     }
     response = requests.get(
-        url,
+        url=url,
         headers=headers,
     )
+
     logging.info(f"""
+    Gets all enrollments.
+    
     Sending request to {response.url}
     Response: {response.text}
     """)
+
     if response.status_code == 200:
         return {
             "status": "success",
@@ -66,18 +86,28 @@ def get_enrollment_by_id(base_domain, token, enrollment_id):
     """
     Gets an enrollment by its ID.
     """
+    if not token:
+        return {
+            'status': 'error',
+            'message': 'Token is required'
+        }
+
     url = base_domain + "/enrollments/" + enrollment_id
     headers = {
         "Authorization": "Bearer " + token
     }
     response = requests.get(
-        url,
+        url=url,
         headers=headers,
     )
+
     logging.info(f"""
+    Gets an enrollment by its ID.
+    
     Sending request to {response.url}
     Response: {response.text}
     """)
+
     if response.status_code == 200:
         return {
             "status": "success",
@@ -93,18 +123,28 @@ def delete_enrollment_by_id(base_domain, token, enrollment_id):
     """
     Deletes an enrollment by its ID.
     """
+    if not token:
+        return {
+            'status': 'error',
+            'message': 'Token is required'
+        }
+
     url = base_domain + "/enrollments/" + enrollment_id + "/"
     headers = {
-        "Authorization": "Bearer " + token
+        "Authorization": "Bearer " + token,
     }
     response = requests.delete(
-        url,
+        url=url,
         headers=headers,
     )
+
     logging.info(f"""
+    Deletes an enrollment by its ID.
+     
     Sending request to {response.url}
     Response: {response.text}
     """)
+
     if response.status_code == 204:
         return {
             "status": "success",
